Fix compilation warnings
git-svn-id: https://yabause.svn.sourceforge.net/svnroot/yabause/trunk@1986 d7be59e3-070e-4b9a-85a1-4646befae34d

+++ yabause/src/perlinuxjoy.c
@@ -20,6 +20,7 @@
 #include "debug.h"
 #include "perlinuxjoy.h"
 #include <linux/joystick.h>
+#include <unistd.h>
 #include <fcntl.h>
 #include <errno.h>
 
@@ -123,5 +124,5 @@ void PERLinuxJoyFlush(void) {
 
 void PERLinuxKeyName(u32 key, char * name, UNUSED int size)
 {
-   sprintf(name, "%x", key);
+   sprintf(name, "%x", (int)key);
 }
